requirements.txt
# Distributed Log Processing System - Requirements

# PySpark
pyspark>=3.5.0

# Data Processing (for dashboard only - NOT for Spark processing)
pandas>=2.0.0
pyarrow>=14.0.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Configuration
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.2
